refactor: 계산식의 숫자 예외를 Long에서 Double로 변경

+++ src/main/java/calculator/ui/CheckInputException.java
@@ -45,7 +45,7 @@ public class CheckInputException {
 
     private static void checkNumber(String data) {
         try {
-            Long.parseLong(data);
+            Double.parseDouble(data);
         } catch (NumberFormatException exception) {
             throw new NotEquationFormatException(exception.getMessage());
         }
